Revert oversimplification of DigProvider
Unfortunately, dig does not correctly handle the specification of multiple
`+domain` arguments, discarding all but the last one.  Therefore we need
to run it multiple times and combine the results if multiple
`search_domains` are specified.

This is a very important use case for me, and several other known users.