Explain what user is doing wrong if they try to use OpenConnect with both '--script-tun' and '--script vpn-slice'
When OpenConnect is running in its transparent proxy mode, with
`--script-tun`, it can't be used with vpn-slice (nor the standard
vpnc-script, for that matter).  As the documentation explains, either
'ocproxy' or 'tunsocks' is needed in this case:
https://openconnect.gitlab.io/openconnect/nonroot.html

Some vpn-slice users have been confused by this.  We can use the fact that
OpenConnect sets the environment variable VPNFD in transparent proxy mode
to detect this condition and explain what's wrong:
https://gitlab.com/openconnect/openconnect/blob/0621ef940d28a95e3aa9054e0c02f6c7ab1153d3/tun.c#L499

Fixes #70