Add '--kerberos-dc REALM' option
In https://github.com/dlenski/vpn-slice/issues/135, I learned that it's
possible to automatically lookup the Kerberos5 domain controller (DC) hosts
for a given Kerberos5 realm, by using the DNS SRV records.

For example:

    $ REALM=realm.bigcorp.com
    $ dig +short @dns0.vpn0 @dns1.vpn0 _kerberos._tcp.$REALM SRV
    0 100 88 dc-01.realm.bigcorp.com.
    0 100 88 dc-02.realm.bigcorp.com.

This makes it quite easy to automatically lookup and route the DC hosts in
vpn-slice, using most of the existing machinery for looking up
user-specified hostnames.